@@ -4,37 +4,44 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^6.3.0",
-    "@emotion/babel-preset-css-prop": "^11.2.0",
-    "@emotion/react": "^11.4.0",
-    "@emotion/styled": "^11.3.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.15",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "axios": "^0.21.4",
-    "craco-less": "^1.20.0",
-    "i18next": "^21.2.0",
-    "i18next-browser-languagedetector": "^6.1.2",
+    "@emotion/babel-preset-css-prop": "^11.10.0",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "assert": "^2.1.0",
+    "axios": "^1.2.6",
+    "file-loader": "^6.2.0",
+    "i18next": "^22.4.9",
+    "i18next-browser-languagedetector": "^7.0.1",
     "ics": "^2.31.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-i18next": "^11.12.0",
-    "react-markdown": "^7.0.1",
-    "react-router": "^5.2.1",
-    "react-router-dom": "^5.3.0",
+    "react-i18next": "^12.1.4",
+    "react-markdown": "^8.0.5",
+    "react-quiz-component": "^0.5.1",
+    "react-router": "^6.8.0",
+    "react-router-dom": "^6.8.0",
     "react-router-hash-link": "^2.4.3",
-    "react-scripts": "4.0.3",
-    "remark-gfm": "^2.0.0",
-    "typescript": "^4.1.2",
-    "web-vitals": "^1.0.1"
+    "remark-gfm": "^3.0.1",
+    "typescript": "^4.9.4",
+    "web-vitals": "^3.1.1"
+  },
+  "devDependencies": {
+    "@craco/craco": "7.0.0",
+    "@svgr/webpack": "^6.5.1",
+    "@types/jest": "^29.4.0",
+    "@types/node": "^18.11.18",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.10",
+    "@types/react-helmet": "^6.1.7",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-router-hash-link": "^2.4.5",
+    "craco-less": "2.1.0-alpha.0",
+    "react-scripts": "^5.0.1",
+    "serve": "^14.2.0"
   },
   "scripts": {
     "start": "PORT=3001 craco start",
@@ -59,10 +66,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@types/react-helmet": "^6.1.1",
-    "@types/react-router-dom": "^5.1.9",
-    "@types/react-router-hash-link": "^2.4.2"
   }
 }
